+++ geolocator_web/lib/src/utils.dart
@@ -18,8 +18,10 @@ Position toPosition(html.Geoposition webPosition) {
         ? DateTime.fromMillisecondsSinceEpoch(webPosition.timestamp!)
         : DateTime.now(),
     altitude: coords.altitude as double? ?? 0.0,
+    altitudeAccuracy: coords.altitudeAccuracy as double? ?? 0.0,
     accuracy: coords.accuracy as double? ?? 0.0,
     heading: coords.heading as double? ?? 0.0,
+    headingAccuracy: 0.0,
     floor: null,
     speed: coords.speed as double? ?? 0.0,
     speedAccuracy: 0.0,