+++ geolocator_web/test/geolocator_web_test.dart
@@ -15,8 +15,10 @@ List<Position> get mockPositions => List.of(() sync* {
               isUtc: true,
             ),
             altitude: 3000.0,
+            altitudeAccuracy: 0.0,
             accuracy: 0.0,
             heading: 0.0,
+            headingAccuracy: 0.0,
             speed: 0.0,
             speedAccuracy: 0.0);
       }
@@ -150,8 +152,10 @@ void main() {
                 isUtc: true,
               ),
               altitude: 3000.0,
+              altitudeAccuracy: 0.0,
               accuracy: 0.0,
               heading: 0.0,
+              headingAccuracy: 0.0,
               speed: 0.0,
               speedAccuracy: 0.0);
         }