+++ geolocator_linux/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.0
+
+- Bumps `geolocator_platform_interface` dependency to `^4.1.0`.
+
 ## 0.1.3
 
 - Improves GNOME detection.